backend/requirements.txt
# Backend
fastapi
uvicorn
langchain-google-genai  
langgraph
rdkit                   
python-dotenv           

# Infrastructure
pydantic
typing-extensions
